adding index definitions and marshalling/unmarshalling funcs
