+++ lib/recipes-data/.eslintrc.json
@@ -1,6 +1,7 @@
 {
 	"extends": ["../../.eslintrc.json"],
 	"ignorePatterns": ["!**/*"],
+
 	"overrides": [
 		{
 			"files": ["*.ts", "*.tsx", "*.js", "*.jsx"],
@@ -8,6 +9,9 @@
 		},
 		{
 			"files": ["*.ts", "*.tsx"],
+      "parserOptions": {
+        "project": ["lib/recipes-data/tsconfig.lib.json"]
+      },
 			"rules": {}
 		},
 		{
