--- lib/recipes-data/src/lib/lib-recipes-data.spec.ts
@@ -1,7 +1,0 @@
-import { libRecipesData } from './lib-recipes-data';
-
-describe('libRecipesData', () => {
-	it('should work', () => {
-		expect(libRecipesData()).toEqual('lib-recipes-data');
-	});
-});
