--- lib/recipes-data/src/lib/lib-recipes-data.ts
@@ -1,3 +1,0 @@
-export function libRecipesData(): string {
-	return 'lib-recipes-data';
-}
